requirements.txt
aiohttp==3.8.5
aiomysql==0.2.0
aiosignal==1.3.1
aiosqlite==0.17.0
attrs==23.2.0
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
Flask==2.3.4          # Совместим с Werkzeug 2.3.6
frozenlist==1.4.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
multidict==6.0.5
openai==1.48     # Обновлено для поддержки новых функций и исправлений
PyMySQL==1.1.1
python-dotenv==0.21.0
requests==2.31.0
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.2        # Обновлено для улучшения совместимости и безопасности
Werkzeug==2.3.6       # Совместим с Flask 2.3.4
yarl==1.9.4
pandas==2.0.5
scikit-learn==1.3.2
nltk==3.6.7
setuptools>=58.0.0
mosaicml==0.23.5
torch==2.1.2
requests_mock==1.9.3
sqlalchemy==2.0.16
prompt-toolkit==3.0.41
babel==2.10
numpy==1.24.4
quart==0.19.6
theano==1.0.5
